requirements.txt
alembic==1.13.3
annotated-types==0.7.0
colorama==0.4.6
greenlet==3.1.1
iniconfig==2.0.0
Jinja2==3.1.4
Mako==1.3.5
MarkupSafe==2.1.5
packaging==24.1
pluggy==1.5.0
pydantic==2.9.2
pydantic_core==2.23.4
PyMySQL==1.1.1
pytest==8.3.3
python-dotenv==1.0.1
SQLAlchemy==2.0.35
typing_extensions==4.12.2
waitress==3.0.0
whitenoise==6.7.0
